rename: buttonWithIcon to ButtonWithDropdown to avoid duplication
