+++ frontend/src/components/UI/button/ButtonWithDropdown.tsx
@@ -3,25 +3,25 @@ import { ReactComponent as ChevronDownIcon } from "../../../assets/icons/chevron
 import { Fragment } from "react";
 import { Menu, Transition } from "@headlessui/react";
 
-const ButtonWithIcon: React.FC<{
+const ButtonWithDropdown: React.FC<{
 	icon: JSX.Element;
 	text: string;
 	dropDown: boolean;
 	options?: string[];
 }> = (props) => {
 	return (
-		<div className="flex items-center px-5 py-3 text-lg bg-dark-60 rounded-2xl gap-1">
+		<div className="flex items-center gap-1 px-5 py-3 text-lg bg-dark-60 rounded-2xl">
 			<Menu as="div" className="relative inline-block text-left">
 				<div>
-					<Menu.Button className="flex flex-row items-center gap-2 w-full justify-center rounded-md border border-dark-60 hover:border-white bg-dark-60 px-4 py-2  focus:outline-none">
+					<Menu.Button className="flex flex-row items-center justify-center w-full gap-2 px-4 py-2 border rounded-md border-dark-60 hover:border-white bg-dark-60 focus:outline-none">
 						{props.icon}
-						<p className="text-white font-medium text-sm">
+						<p className="text-sm font-medium text-white">
 							{props.text}
 						</p>
 
 						{props.dropDown && (
 							<ChevronDownIcon
-								className="h-4 w-4"
+								className="w-4 h-4"
 								aria-hidden="true"
 							/>
 						)}
@@ -38,12 +38,12 @@ const ButtonWithIcon: React.FC<{
 						leaveFrom="transform opacity-100 scale-100"
 						leaveTo="transform opacity-0 scale-95"
 					>
-						<Menu.Items className="absolute right-0 z-10 mt-2 w-56 origin-top-right rounded-md bg-dark-60 shadow-lg ring-1 ring-white focus:outline-none">
+						<Menu.Items className="absolute right-0 z-10 w-56 mt-2 origin-top-right rounded-md shadow-lg bg-dark-60 ring-1 ring-white focus:outline-none">
 							<div className="p-2">
 								{props.options.map((option, index) => {
 									return (
 										<div
-											className="hover:text-dark-grey cursor-pointer	"
+											className="cursor-pointer hover:text-dark-grey "
 											key={index}
 										>
 											<Menu.Item>
@@ -68,4 +68,4 @@ const ButtonWithIcon: React.FC<{
 	);
 };
 
-export default ButtonWithIcon;
+export default ButtonWithDropdown;
